backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
pyjwt
pydantic[email]
